WOMA-66: Apply naming conventions from recipe categories to ingredients
