+++ core/src/zeit/wochenmarkt/browser/categories_widget.pt
@@ -1,4 +1,4 @@
-<div class="categories-widget"
+<div class="recipe-categories-widget"
   i18n:domain="zeit.cms"
   tal:attributes="id string:${view/name}.wrapper">
 
@@ -14,7 +14,7 @@
 </ol>
 
 <input
- type="text" name="add_category" placeholder="Type to find recipe categories ..."
+ type="text" name="add_recipe_category" placeholder="Type to find recipe categories ..."
  i18n:attributes="placeholder"
  tal:attributes="cms:autocomplete-source view/autocomplete_source_url;
                  id string:${view/name}.add"
