+++ core/src/zeit/wochenmarkt/tests/test_ingredients.py
@@ -10,25 +10,25 @@ class TestIngredients(zeit.wochenmarkt.testing.FunctionalTestCase):
 
     def test_ingredient_should_be_found_through_xml(self):
         ingredients = zope.component.getUtility(
-            zeit.wochenmarkt.interfaces.IIngredients)._load()
+            zeit.wochenmarkt.interfaces.IIngredientsWhitelist)._load()
         basmati = dict(ingredients.items()).get('basmatireis')
         assert 'Basmatireis' == basmati.name
         assert 'other' == basmati.category
 
     def test_ingredient_should_be_found_by_id(self):
         calamari = zope.component.getUtility(
-            zeit.wochenmarkt.interfaces.IIngredients).get('calamari')
+            zeit.wochenmarkt.interfaces.IIngredientsWhitelist).get('calamari')
         assert 'Calamari' == calamari.name
         assert 'fish' == calamari.category
 
     def test_ingredients_should_be_found_by_category(self):
         meat_ingredients = zope.component.getUtility(
-            zeit.wochenmarkt.interfaces.IIngredients).category('meat')
+            zeit.wochenmarkt.interfaces.IIngredientsWhitelist).category('meat')
         assert 2 == len(meat_ingredients)
 
     def test_autocomplete_should_be_available_for_ingredients(self):
         result = zope.component.getUtility(
-            zeit.wochenmarkt.interfaces.IIngredients).search('B')
+            zeit.wochenmarkt.interfaces.IIngredientsWhitelist).search('B')
         assert 6 == len(result)
         names = []
         for item in result:
@@ -37,6 +37,7 @@ class TestIngredients(zeit.wochenmarkt.testing.FunctionalTestCase):
 
     def test_ingredients_should_be_found_through_multiple_criteria(self):
         chicken = zope.component.getUtility(
-            zeit.wochenmarkt.interfaces.IIngredients).category('chicken', 'B')
+            zeit.wochenmarkt.interfaces.IIngredientsWhitelist).category(
+                'chicken', 'B')
         assert 1 == len(chicken)
         assert u'Brathähnchen' == chicken[0].name
